Exception handling - Delete

+++ src/main/java/com/yrtweb/course/service/UserService.java
@@ -2,14 +2,16 @@ package com.yrtweb.course.service;
 
 import com.yrtweb.course.entities.User;
 import com.yrtweb.course.repositories.UserRepository;
+import com.yrtweb.course.service.exceptions.DatabaseException;
 import com.yrtweb.course.service.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.webmvc.autoconfigure.WebMvcProperties;
-import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Repository;
+
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
+
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -29,12 +31,22 @@ public class UserService {
     }
 
     public User insert (User obj){
-       return ur.save(obj);
+
+        return ur.save(obj);
     }
 
     public void delete(Long id){
-        ur.deleteById(id);
+        if (!ur.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+        try{
+            ur.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+          throw new DatabaseException(e.getMessage());
+        }
     }
+
     public User update(Long id , User obj){
         User entity = ur.getReferenceById(id);
         updateData(entity,obj);
